requirements.txt
flask>=2.3.0
psycopg2-binary>=2.9.0
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
yfinance>=0.2.0
yahooquery>=2.3.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.0
jinja2>=3.0.0
python-dotenv>=0.19.0
gunicorn>=21.0.0
openpyxl>=3.0.0
flask
gunicorn
matplotlib
numpy
openpyxl
pandas
psycopg2-binary
python-dateutil
